+++ guapow/service/optimizer/task/model.py
@@ -4,7 +4,7 @@ from typing import Optional, Tuple, Dict, Set, Type
 
 from guapow.common.dto import OptimizationRequest
 from guapow.common.model import ScriptSettings
-from guapow.service.optimizer.cpu import CPUFrequencyManager
+from guapow.service.optimizer.cpu import CPUFrequencyManager, CPUEnergyPolicyManager
 from guapow.service.optimizer.flow import OptimizationQueue
 from guapow.service.optimizer.gpu import GPUManager, GPUDriver, GPUState
 from guapow.service.optimizer.mouse import MouseCursorManager
@@ -14,15 +14,18 @@ from guapow.service.optimizer.win_compositor import WindowCompositor
 
 class OptimizationContext:
 
-    def __init__(self, gpu_man: Optional[GPUManager], logger: Optional[Logger], cpufreq_man: Optional[CPUFrequencyManager],
+    def __init__(self, gpu_man: Optional[GPUManager], logger: Optional[Logger],
+                 cpufreq_man: Optional[CPUFrequencyManager], cpuenergy_man: Optional[CPUEnergyPolicyManager],
                  mouse_man: Optional[MouseCursorManager], queue: Optional[OptimizationQueue], cpu_count: int,
-                 launcher_mapping_timeout: float, renicer_interval: float, compositor: Optional[WindowCompositor] = None,
-                 allow_root_scripts: bool = False, compositor_disabled_context: Optional[dict] = None):
+                 launcher_mapping_timeout: float, renicer_interval: float,
+                 compositor: Optional[WindowCompositor] = None,  allow_root_scripts: bool = False,
+                 compositor_disabled_context: Optional[dict] = None):
 
         self.queue = queue
         self.gpu_man = gpu_man
         self.logger = logger
         self.cpufreq_man = cpufreq_man
+        self.cpuenergy_man = cpuenergy_man
         self.mouse_man = mouse_man
         self.cpu_count = cpu_count
         self.compositor = compositor
@@ -34,7 +37,7 @@ class OptimizationContext:
     @classmethod
     def empty(cls) -> "OptimizationContext":
         return cls(gpu_man=None, mouse_man=None, logger=None, cpufreq_man=None, queue=None,
-                   cpu_count=0, launcher_mapping_timeout=0, renicer_interval=0)
+                   cpu_count=0, launcher_mapping_timeout=0, renicer_interval=0, cpuenergy_man=None)
 
     async def is_mouse_cursor_hidden(self) -> Optional[bool]:
         return await self.mouse_man.is_cursor_hidden() if self.mouse_man else None
@@ -61,8 +64,9 @@ class CPUState:
 class OptimizedProcess:
 
     def __init__(self, request: OptimizationRequest, created_at: float, profile: Optional[OptimizationProfile] = None,
-                 previous_gpus_states: Optional[Dict[Type[GPUDriver], Set[GPUState]]] = None, previous_cpu_state: Optional[CPUState] = None,
-                 stopped_after_launch: Optional[Dict[str, str]] = None):
+                 previous_gpus_states: Optional[Dict[Type[GPUDriver], Set[GPUState]]] = None,
+                 previous_cpu_state: Optional[CPUState] = None, stopped_after_launch: Optional[Dict[str, str]] = None,
+                 cpu_energy_policy_changed: bool = False):
         self.created_at = created_at
         self.request = request
         self.profile = profile
@@ -72,6 +76,7 @@ class OptimizedProcess:
         self.alive = True
         self.related_pids = {*self.request.related_pids} if self.request and self.request.related_pids else set()
         self.pid = request.pid if self.request else None
+        self.cpu_energy_policy_changed = cpu_energy_policy_changed
 
     def should_be_watched(self) -> bool:
         return bool(self.pid is not None and any([self.related_pids,
@@ -81,7 +86,8 @@ class OptimizedProcess:
                                                   self.requires_compositor_disabled,
                                                   self.stopped_processes,
                                                   self.requires_mouse_hidden,
-                                                  self.stopped_after_launch]))
+                                                  self.stopped_after_launch,
+                                                  self.cpu_energy_policy_changed]))
 
     @property
     def source_pid(self) -> Optional[int]: